refactor: one header per pass category, one copy of the VM opcode table
Passes.h was 400 lines whose section comments had stopped matching the
directory layout — Substitution and CSEBreak filed under "Control Flow" with
sources in Data/, VTableProtection under "RTTI / vtable protection" with its
source in CFG/. Each declaration now sits in include/kagura/Passes/<Category>.h
named for the lib/Transforms/ subdirectory that implements it, and every pass
source includes only its own category. Passes.h stays as the umbrella for
Plugin.cpp and the fuzz targets, so nothing outside had to change.

ObfuscationMetrics.cpp moves from the top level into Infrastructure/, where the
other non-obfuscating passes live. lib/Transforms/ top level is now exactly the
three files that are not passes.

The VM opcode table was two hand-written copies of a wire format: a C++ enum in
include/kagura/VM.h and 48 #defines in runtime/core/vm_interpreter.c, plus four
frame limits, with a comment in each file reminding the reader to change the
other. A disagreement between them does not fail to compile, does not fail a
test that inspects IR, and shows up as a shipped binary computing the wrong
thing. They agreed today; nothing kept them agreeing.

Both sides now expand include/kagura/VMOpcodes.def, which also carries the
per-opcode operand documentation, so the ISA is described once. The interpreter
gets enums rather than #defines while we are here, so a debugger can name the
opcode it is stopped on.

That means the C runtime now includes a header out of include/, which needed
three build files taught about it: Package.swift had no header search path at
all, kagura.bzl globbed only include/**/*.h into hdrs (which also left
runtime/internal.h undeclared — fixed too), and the podspec listed source_files
file by file.

The install rule needed fixing regardless. It globbed include/kagura with
FILES_MATCHING PATTERN "*.h", which shipped the LLVM pass-plugin headers to
consumers who have no LLVM, and once those headers started expanding a .def the
pattern excluded it — so the installed Options.h could not be included at all.
It now lists the three files that are actually the shipped interface, drawing
the same line the podspec and Package.swift already documented.

Verified by compiling a consumer against `cmake --install` output.
