+++ lib/Transforms/AntiAnalysis/AntiTamper.cpp
@@ -40,7 +40,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "kagura/Options.h"
-#include "kagura/Passes.h"
+#include "kagura/Passes/AntiAnalysis.h"
 #include "kagura/Utils.h"
 
 #include "llvm/IR/BasicBlock.h"
