+++ lib/Transforms/AntiAnalysis/HoneyValue.cpp
@@ -25,7 +25,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "kagura/Options.h"
-#include "kagura/Passes.h"
+#include "kagura/Passes/AntiAnalysis.h"
 #include "kagura/Utils.h"
 
 #include "llvm/IR/Constants.h"
