+++ lib/Transforms/AntiAnalysis/SymbolVisibility.cpp
@@ -31,7 +31,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "kagura/Options.h"
-#include "kagura/Passes.h"
+#include "kagura/Passes/AntiAnalysis.h"
 #include "kagura/Utils.h"
 
 #include "llvm/IR/Function.h"
