+++ lib/Transforms/CFG/BasicBlockSplitting.cpp
@@ -26,7 +26,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "kagura/Options.h"
-#include "kagura/Passes.h"
+#include "kagura/Passes/CFG.h"
 #include "kagura/Utils.h"
 
 #include "llvm/IR/BasicBlock.h"
