+++ lib/Transforms/CFG/DeadCodeInsertion.cpp
@@ -19,7 +19,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "kagura/Options.h"
-#include "kagura/Passes.h"
+#include "kagura/Passes/CFG.h"
 #include "kagura/Utils.h"
 
 #include "llvm/IR/BasicBlock.h"
