+++ lib/Transforms/CFG/LoopTransform.cpp
@@ -20,7 +20,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "kagura/Options.h"
-#include "kagura/Passes.h"
+#include "kagura/Passes/CFG.h"
 #include "kagura/Utils.h"
 
 #include "llvm/Analysis/LoopInfo.h"
