+++ lib/Transforms/Data/ConstantObfuscation.cpp
@@ -25,7 +25,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "kagura/Passes.h"
+#include "kagura/Passes/Data.h"
 #include "kagura/Utils.h"
 
 #include "llvm/IR/Constants.h"
