+++ lib/Transforms/Data/EncryptedLookupTable.cpp
@@ -37,7 +37,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "kagura/Options.h"
-#include "kagura/Passes.h"
+#include "kagura/Passes/Data.h"
 #include "kagura/Utils.h"
 
 #include "llvm/ADT/ArrayRef.h"
