+++ lib/Transforms/Data/GlobalEncryption.cpp
@@ -11,7 +11,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "kagura/Options.h"
-#include "kagura/Passes.h"
+#include "kagura/Passes/Data.h"
 #include "kagura/Utils.h"
 
 #include "llvm/IR/Constants.h"
