+++ lib/Transforms/Data/MemoryValueObfuscation.cpp
@@ -26,7 +26,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "kagura/Options.h"
-#include "kagura/Passes.h"
+#include "kagura/Passes/Data.h"
 #include "kagura/Utils.h"
 
 #include "llvm/IR/Function.h"
