+++ lib/Transforms/Data/StringEncryption.cpp
@@ -26,7 +26,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "kagura/Passes.h"
+#include "kagura/Passes/Data.h"
 #include "kagura/Utils.h"
 
 #include "llvm/IR/Constants.h"
