+++ lib/Transforms/Data/StringSplit.cpp
@@ -64,7 +64,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "kagura/Passes.h"
+#include "kagura/Passes/Data.h"
 #include "kagura/Utils.h"
 
 #include "llvm/IR/Constants.h"
