+++ lib/Transforms/Infrastructure/AuditLog.cpp
@@ -21,7 +21,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "kagura/Options.h"
-#include "kagura/Passes.h"
+#include "kagura/Passes/Infrastructure.h"
 #include "kagura/Utils.h"
 
 #include "llvm/IR/DebugInfoMetadata.h"
