+++ lib/Transforms/Infrastructure/ConfigLoader.cpp
@@ -49,7 +49,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "kagura/Options.h"
-#include "kagura/Passes.h"
+#include "kagura/Passes/Infrastructure.h"
 
 #include "llvm/IR/Module.h"
 #include "llvm/Support/JSON.h"
