+++ lib/Transforms/Platform/JNIObfuscation.cpp
@@ -17,7 +17,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "kagura/Passes.h"
+#include "kagura/Passes/Platform.h"
 #include "kagura/Utils.h"
 
 #include "llvm/IR/Constants.h"
