+++ lib/Transforms/Platform/ObjCObfuscation.cpp
@@ -32,7 +32,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "kagura/Passes.h"
+#include "kagura/Passes/Platform.h"
 #include "kagura/Utils.h"
 
 #include "llvm/IR/BasicBlock.h"
